refactor: split up pkg flow with normal node flow
