+++ utils/index.ts
@@ -2,3 +2,4 @@ export * from './distanceTo';
 export * from './isInLicenseRange';
 export * from './enums';
 export * from './retry';
+export * from './cliValidators';
